+++ src/main/java/com/care/dao/JobDAO.java
@@ -10,16 +10,16 @@ public interface JobDAO extends DAO {
     int addJob(Job job) throws SQLException;
 
     // deletion is nothing but marking the status flag as INACTIVE
-    int setJobStatus(int jobId, Status status) throws SQLException;
+    int setJobStatus(long jobId, Status status) throws SQLException;
 
-    int setAllJobsStatus(int postedBy, Status status) throws SQLException;
+    int setAllJobsStatus(long postedBy, Status status) throws SQLException;
 
     int editJob(Job jobId) throws SQLException;
 
-    Job getJob(int jobId) throws SQLException;
+    Job getJob(long jobId) throws SQLException;
 
-    List<Job> getAllJobs(int postedBy) throws SQLException;
+    List<Job> getAllJobs(long postedBy) throws SQLException;
 
-    List<Job> getAllAvailableJobs(int sitterId)throws SQLException;
+    List<Job> getAllAvailableJobs(long sitterId)throws SQLException;
 
 }
